refactor: Move McStopArrivals for c1 to package c1
