+++ src/main/java/org/opentripplanner/raptor/rangeraptor/multicriteria/McRangeRaptorWorkerState.java
@@ -9,10 +9,10 @@ import org.opentripplanner.raptor.api.model.RaptorTripSchedule;
 import org.opentripplanner.raptor.rangeraptor.internalapi.RaptorWorkerResult;
 import org.opentripplanner.raptor.rangeraptor.internalapi.RaptorWorkerState;
 import org.opentripplanner.raptor.rangeraptor.internalapi.WorkerLifeCycle;
-import org.opentripplanner.raptor.rangeraptor.multicriteria.arrivals.AccessStopArrival;
 import org.opentripplanner.raptor.rangeraptor.multicriteria.arrivals.McStopArrival;
-import org.opentripplanner.raptor.rangeraptor.multicriteria.arrivals.TransferStopArrival;
-import org.opentripplanner.raptor.rangeraptor.multicriteria.arrivals.TransitStopArrival;
+import org.opentripplanner.raptor.rangeraptor.multicriteria.arrivals.c1.AccessStopArrival;
+import org.opentripplanner.raptor.rangeraptor.multicriteria.arrivals.c1.TransferStopArrival;
+import org.opentripplanner.raptor.rangeraptor.multicriteria.arrivals.c1.TransitStopArrival;
 import org.opentripplanner.raptor.rangeraptor.multicriteria.heuristic.HeuristicsProvider;
 import org.opentripplanner.raptor.rangeraptor.path.DestinationArrivalPaths;
 import org.opentripplanner.raptor.rangeraptor.transit.RaptorTransitCalculator;
